Change data type to long

+++ DScannerLibrary/Models/ArticleModel.cs
@@ -6,7 +6,7 @@ public class ArticleModel
     public string? denumire { get; set; }
     public decimal tva { get; set; }
     public decimal pret_vanz { get; set; }
-    public int cod_bare { get; set; }
+    public long cod_bare { get; set; }
     //public string? lot { get; set; }
     public string? data_intr { get; set; }
 }
